requirements.txt
selenium
streamlit
pandas
matplotlib
mysql-connector-python